# app/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal World Model API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/connections": {
      "get": {
        "summary": "Connections",
        "operationId": "connections_connections_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionsView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/connections/demo": {
      "post": {
        "summary": "Connect Demo",
        "operationId": "connect_demo_connections_demo_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/connections/{connector}": {
      "delete": {
        "summary": "Disconnect",
        "operationId": "disconnect_connections__connector__delete",
        "parameters": [
          {
            "name": "connector",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connector"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Disconnect Connections  Connector  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me": {
      "delete": {
        "summary": "Delete Everything",
        "description": "Delete the user and, by cascade, everything known about them, and withdraw our access\nat Google. A signed-in account is simply gone; the local development identity starts\nagain empty, because that is all \"signing up again\" means for it.",
        "operationId": "delete_everything_me_delete",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deleted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Available": {
        "properties": {
          "connector": {
            "type": "string",
            "title": "Connector"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "connector",
          "label",
          "available",
          "note"
        ],
        "title": "Available"
      },
...
      "ConnectionView": {
        "properties": {
          "connector": {
            "type": "string",
            "title": "Connector"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "connected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Connected At"
          },
          "last_synced_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Synced At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "sources": {
            "type": "integer",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "connector",
          "label",
          "connected_at",
          "last_synced_at",
          "status",
          "sources"
        ],
        "title": "ConnectionView"
      },
      "ConnectionsView": {
        "properties": {
          "connected": {
            "items": {
              "$ref": "#/components/schemas/ConnectionView"
            },
            "type": "array",
            "title": "Connected"
          },
          "available": {
            "items": {
              "$ref": "#/components/schemas/Available"
            },
            "type": "array",
            "title": "Available"
          },
          "understood": {
            "type": "integer",
            "title": "Understood"
          }
        },
        "type": "object",
        "required": [
          "connected",
          "available",
          "understood"
        ],
        "title": "ConnectionsView"
      },
...
      "Deleted": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "google_access_revoked": {
            "type": "boolean",
            "title": "Google Access Revoked"
          }
        },
        "type": "object",
        "required": [
          "status",
          "google_access_revoked"
        ],
        "title": "Deleted"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SyncResult": {
        "properties": {
          "new_sources": {
            "type": "integer",
            "title": "New Sources"
          },
          "understood": {
            "type": "integer",
            "title": "Understood"
          }
        },
        "type": "object",
        "required": [
          "new_sources",
          "understood"
        ],
        "title": "SyncResult"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}